@@ -11,34 +11,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.13",
-    "@angular/cdk": "~8.2.3",
-    "@angular/common": "~8.2.13",
-    "@angular/compiler": "~8.2.13",
-    "@angular/core": "~8.2.13",
-    "@angular/forms": "~8.2.13",
+    "@angular/animations": "~10.0.4",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "~10.0.4",
+    "@angular/compiler": "~10.0.4",
+    "@angular/core": "~10.0.4",
+    "@angular/forms": "~10.0.4",
     "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "~8.2.13",
-    "@angular/platform-browser-dynamic": "~8.2.13",
-    "@angular/router": "~8.2.13",
+    "@angular/platform-browser": "~10.0.4",
+    "@angular/platform-browser-dynamic": "~10.0.4",
+    "@angular/router": "~10.0.4",
     "@fortawesome/angular-fontawesome": "^0.6.1",
+    "@fortawesome/fontawesome-free": "^5.13.1",
+    "@m0t0r/ngx-slider": "^1.0.0",
+    "@types/chart.js": "^2.9.22",
     "angular-font-awesome": "^3.1.2",
+    "animate.css": "^4.1.0",
+    "blueimp-canvas-to-blob": "^3.28.0",
     "bootstrap": "^4.4.1",
+    "chart.js": "^2.9.3",
+    "detect-passive-events": "^1.0.4",
     "font-awesome": "^4.7.0",
+    "hammerjs": "^2.0.8",
     "material-design-icons": "^3.0.1",
+    "ng2-img-max": "^2.2.4",
+    "ng5-slider": "^1.2.4",
     "ngx-bootstrap": "^5.6.1",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "rxjs": "~6.6.0",
+    "rxjs-compat": "^6.6.0",
+    "tslib": "^1.13.0",
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.16",
-    "@angular/cli": "~8.3.16",
-    "@angular/compiler-cli": "~8.2.13",
-    "@angular/language-service": "~8.2.13",
-    "@types/node": "~8.9.4",
+    "@angular-devkit/build-angular": "^0.1000.3",
+    "@angular/cli": "^10.0.3",
+    "@angular/compiler-cli": "~10.0.4",
+    "@angular/language-service": "~10.0.4",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
     "codelyzer": "^5.0.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -50,6 +61,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "typescript": "~3.9.6"
   }
 }
