@@ -1,8 +1,10 @@
 export class Grad {
   ptt: number;
   naziv: string;
-  constructor(ptt, naziv) {
+  brojKorisnika: number;
+  constructor(ptt, naziv, brojKorisnika?: number) {
     this.ptt = ptt;
     this.naziv = naziv;
+    this.brojKorisnika = brojKorisnika;
   }
 }
