@@ -4,9 +4,11 @@ export class Kategorija {
   id: number;
   naziv: string;
   slika?: Slika;
-  constructor(id, naziv, slika?) {
+  brojArtikala?: number
+  constructor(id, naziv, slika?, brojArtikala?: number) {
     this.id = id;
     this.naziv = naziv;
     this.slika = slika;
+    this.brojArtikala = brojArtikala;
   }
 }
