@@ -10,8 +10,9 @@ export class Prodavac {
   ulica: string;
   gazdinstvo?: string;
   opis?: string;
+  id?: number;
 
-  constructor(ime: string, prezime: string, email: string, lozinka: string, kontakt: string, grad: Grad, ulica: string, gazdinstvo?: string, opis?: string) {
+  constructor(ime: string, prezime: string, email: string, lozinka: string, kontakt: string, grad: Grad, ulica: string, gazdinstvo?: string, opis?: string,id? :number) {
       this.ime = ime;
       this.prezime = prezime;
       this.email = email;
@@ -21,5 +22,6 @@ export class Prodavac {
       this.ulica = ulica;
       this.gazdinstvo = gazdinstvo;
       this.opis = opis;
+      this.id = id;
   }
 }
