@@ -4,6 +4,9 @@ import { ApiServiceService } from '../services/api-service.service';
 import { KorpaDto } from '../dtos/korpaDto.model';
 import { AuthService } from '../services/auth.service';
 import { OrderDto } from '../dtos/orderDto.model';
+import { MatDialog } from '@angular/material';
+import { MyDialogComponent } from '../my-dialog/my-dialog.component';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-order',
@@ -18,6 +21,8 @@ export class OrderComponent implements OnInit {
   cities: Grad[];
   selectedCity: Grad;
   korpa: KorpaDto;
+
+
   formImg(picByte){
     return 'data:image/jpeg;base64,'+picByte;
   }
@@ -27,15 +32,16 @@ export class OrderComponent implements OnInit {
     this.address = adresa;
     this.contact = kontaktAlternativni;
   }
-  constructor(private api: ApiServiceService, private auth : AuthService) { }
+  constructor(private api: ApiServiceService, private auth : AuthService, private dialog: MatDialog, private router: Router) { }
 
   ngOnInit() {
     this.api.vratiListuGradova().subscribe(
       response => {
         this.cities = response;
+        this.cities.sort((a, b) => (a.naziv < b.naziv ? -1 : 1));
       }
     );
-    this.korpa = this.auth.getKorpa();
+    this.korpa = this.auth.getCart();
     console.log('Korpa');
     console.log(this.korpa);
   }
@@ -47,13 +53,23 @@ export class OrderComponent implements OnInit {
     stavka.ukupnaVrednost = sum;
     return sum;
   }
-  potvrdi() {
-    var order = new OrderDto(-1, this.korpa.ukupnaVrednost, this.korpa, false, this.auth.userId, null, null, this.address, this.selectedCity, this.contact);
-    console.log('ovo se cuva')
-    console.log(order)
+  potvrdi(firstFormGroup) {
+    this.populateAddress(firstFormGroup);
+    var order = new OrderDto(-1, this.korpa.ukupnaVrednost, this.korpa, false, this.auth.getAuthenticatedUserId(), null, null, this.address, this.selectedCity, this.contact);
     this.api.sacuvajNarudzbine(order).subscribe(
       response => {
-        console.log(response)
+        var status = response['status'];
+        if (status === 'USPESNO') {
+          var poruka = response['poruka'];
+          let dialogRef = this.dialog.open(MyDialogComponent, {data: {status: status, message: poruka}});
+            dialogRef.afterClosed().subscribe(
+              resp => {
+                if (status === 'USPESNO') {
+                  this.router.navigate(['/manage-orders']);
+                }
+              }
+            );
+        }
       }
     )
 
